Move tax info query handlers to usePaymentMethodTaxInfo
