+++ client/my-sites/checkout/composite-checkout/payment-methods/existing-credit-card/types.ts
@@ -1,0 +1,8 @@
+export interface TaxInfo {
+	tax_postal_code: string;
+	tax_country_code: string;
+}
+
+export interface TaxGetInfo extends TaxInfo {
+	is_tax_info_set: boolean;
+}
